Add support for non-async methods that return IEnumerable.
Prompt:

This Roslyn Incremental Source Generator currently only supports methods that return IAsyncEnumerable. Please make it support methods that support IEnumerable as well. See the test case `Emitter_GeneratesCorrectCode_ForIEnumerable` in @CsvGeneratorEmitterTests.cs and the generated output in @IEnumerable.generated.txt for an example of the expected input and output for the source generator.