Simplify indentation managment:
Prompt:

In this Roslyn Incremental Source Generator, the file @CsvGenerator.Emitter.cs manually appends the indenting for every line. Please add a AppendLine to the Emitter class that adds the indentation and `IncreaseIndent` & `DecreaseIndent` methods to increase the level of indentation. Refactor the writing of indentation to use these new methods.